Delete table is almost working, just need to fix the asynchronous updates.

+++ COSC360_M3/php/adminDisplayUsers.php
@@ -52,7 +52,13 @@ if ($pstmt) {
             if(mysqli_num_rows($postResult) > 0) {
                 echo "<tr><td colspan='2'><table class='postTable'><tr><th>Post ID</th><th colspan=2>Content</th></tr>";
                 while ($postRow = mysqli_fetch_assoc($postResult)) {
-                    echo "<tr><td>" . $postRow['PostId'] . "</td><td>" . $postRow['Content'] . "</td><td><button id='delete' name='delete' onclick='deleteRow(this)'>Delete</button></td></tr>";
+
+                    echo "<tr><td>" . $postRow['PostId'] . "</td><td>" . $postRow['Content'] . "</td><td>
+                    <form class='deleteForm' id='deleteForm" . $postRow['PostId'] . "' method='post' action='".$_SERVER["PHP_SELF"]."'>
+                    <input type='hidden' name='delete' value='" . $postRow['PostId'] . "'>
+                    <button type='submit'>Delete</button>
+                    </form></td></tr>";
+                    // echo "<tr><td>" . $postRow['PostId'] . "</td><td>" . $postRow['Content'] . "</td><td><button id='delete' name='delete' onclick='deleteRow(this)'>Delete</button></td></tr>";
                 }
                 echo "</table></td></tr>";
             }
@@ -74,17 +80,33 @@ if ($pstmt) {
 
         }
 
-        // below is the code for deleting a post, but it is not working properly
-        // if delete button is clicked, delete the post from the database 
-        if ($_SERVER["REQUEST_METHOD"] == "POST") {
-            if (isset($_POST['delete'])) {
-                $deletePostId = $_POST['delete'];
-                $deleteSql = "DELETE FROM post WHERE PostId = ?";
-                $deletePstmt = mysqli_prepare($connection, $deleteSql);
-                mysqli_stmt_bind_param($deletePstmt, "s", $deletePostId);
-                mysqli_stmt_execute($deletePstmt);
+        // // below is the code for deleting a post, but it is not working properly
+        // // if delete button is clicked, delete the post from the database 
+        // if ($_SERVER["REQUEST_METHOD"] == "POST") {
+        //     if (isset($_POST['delete'])) {
+        //         $deletePostId = $_POST['delete'];
+        //         $deleteSql = "DELETE FROM post WHERE PostId = ?";
+        //         $deletePstmt = mysqli_prepare($connection, $deleteSql);
+        //         mysqli_stmt_bind_param($deletePstmt, "s", $deletePostId);
+        //         mysqli_stmt_execute($deletePstmt);
+        //     }
+        // }
+
+
+        if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['delete'])) {
+            $deletePostId = $_POST['delete'];
+            $deleteSql = "DELETE FROM post WHERE PostId = ?";
+            $deletePstmt = mysqli_prepare($connection, $deleteSql);
+            mysqli_stmt_bind_param($deletePstmt, "s", $deletePostId);
+            if (mysqli_stmt_execute($deletePstmt)) {
+                echo "Post deleted successfully.";
+                // header("Refresh:0; url=adminMain.php");
+            } else {
+                echo "Error deleting post: " . mysqli_error($connection);
             }
+            exit; // Exit to prevent further processing
         }
+
 
     // No user found
     } else {
